app.py: Sort license entries in `index.*` file by their key
Re-generating the `docs` by utilizing [1] on my machine running Ubuntu
for ScanCode version `31.1.0` produces the same entries for the licenses but
differs in terms of their ordering compared to [1]. This is why I assume
that the ordering is not deterministic, but somehow dependends on the
host system.

Sort the license entries by their license key to ensure a deterministic
order independent of a host system. This leads to nicer diffs for
documentation updates.

[1] `make clean build`
[2] https://github.com/nexB/scancode-licensedb/commit/d29ed2c9b07c5451b2ea6adf91849b07a96caf98

Signed-off-by: Frank Viernau <frank_viernau@epam.com>